requirements.txt
pyttsx3~=2.90
pystray~=0.19.5
Pillow~=10.1.0